+++ logistics_project/apps/tanzania/reminders/__init__.py
@@ -3,14 +3,13 @@ All facilities in the same group get the same reminders on the same schedule.
 All districts get the same reminders on the same schedule.
 
 Reminders can fire up to three times.
- - For all reminders, the second and third are only sent if the previous ones 
+ - For all reminders, the second and third are only sent if the previous ones
    have not been submitted.
  - The text of all three reminders is the same for each category of reminder.
- 
+
 """
 from django.utils.translation import ugettext as _
 from datetime import datetime
-from logistics_project.apps.tanzania.models import SupplyPointStatus
 from threadless_router.router import Router
 from rapidsms import router
 
@@ -24,8 +23,11 @@ def send_reminders(contacts, message):
     for contact in contacts:
         if contact.default_connection and contact.is_active:
             send_message(contact, _(message))
-        
+
 def update_statuses(contacts, type, value):
+    # hack: import happens here to prevent tasks import error
+    from logistics_project.apps.tanzania.models import SupplyPointStatus
+
     now = datetime.utcnow()
     for sp in set(c.supply_point for c in contacts):
         SupplyPointStatus.objects.create(supply_point=sp,
